Tried some sort of IsBusy emulation, let's see if DOAV still works... Fixed a bunch of compiler warnings introduced by previous changes. Added more XB2PC conversion functions, types and constants. IVB buffers now grow when needed, which prevents buffer overflow errors when many SetVertexData4f calls are processed. Disabled the SetRenderState_Deferred patch, as we should be able to pin-point the D3DDeferredRenderState structure reliably across all SDK versions.

+++ src/DxbxKrnl/uEmuKrnlEx.pas
@@ -447,7 +447,7 @@ begin
       InitializeObjectAttributes(@ObjectAttributes, @FileName, OBJ_CASE_INSENSITIVE, 0, NULL);
       Result := xboxkrnl_NtOpenFile(
                 @ConfigPartitionHandle,
-                GENERIC_READ or iif(aIsWriteMode, GENERIC_WRITE, 0) or SYNCHRONIZE,
+                GENERIC_READ or DWORD(iif(aIsWriteMode, GENERIC_WRITE, 0)) or SYNCHRONIZE,
                 @ObjectAttributes,
                 @IoStatusBlock,
                 FILE_SHARE_READ or FILE_SHARE_WRITE,
